+++ src/es_sys_scripts.rs
@@ -24,7 +24,7 @@ pub(crate) fn init_es(rt: &EsRuntimeWrapper) {
 
 fn init_file(runtime: &EsRuntimeWrapper, file_name: &str, es_code: &str) {
     debug!("es_sys_scripts :: evaluating init script {}", file_name);
-    let init_res = runtime.eval_sync(es_code, file_name);
+    let init_res = runtime.eval_void_sync(es_code, file_name);
     if !init_res.is_ok() {
         let esei = init_res.err().unwrap();
         panic!(