+++ src/es_utils/arrays.rs
@@ -4,31 +4,66 @@ use mozjs::conversions::ToJSValConvertible;
 use mozjs::jsapi::IsArray;
 use mozjs::jsapi::JSContext;
 use mozjs::jsapi::JS_GetArrayLength;
+use mozjs::jsapi::JS_GetPropertyById;
+use log::trace;
 
 
 use mozjs::rust::{HandleObject, HandleValue, MutableHandleValue};
+use mozjs::glue::int_to_jsid;
+use crate::es_utils::{EsErrorInfo, report_es_ex};
 
+/// todo, these should return a Result<(something), EsErrorInfo>
 
+/// check whether or not an Object is an Array
 pub fn object_is_array(context: *mut JSContext, obj: HandleObject) -> bool {
     let mut is_array: bool = false;
-    unsafe {
-        IsArray(context, obj.into(), &mut is_array);
-    };
+
+    let ok = unsafe{IsArray(context, obj.into(), &mut is_array)};
+
+    if !ok {
+        if let Some(err) = report_es_ex(context) {
+            trace!("error getting IsArray, ignoring: {}", err.message);
+            return false;
+        }
+    }
     is_array
 }
 
-pub fn get_array_length(context: *mut JSContext, arr_obj: HandleObject) -> u32 {
+/// get the length of an Array
+pub fn get_array_length(context: *mut JSContext, arr_obj: HandleObject) -> Result<u32, EsErrorInfo> {
    let mut l : u32 = 0;
-    unsafe{
-        JS_GetArrayLength(context, arr_obj.into(), &mut l);
-    };
-    l
+
+    let ok = unsafe{JS_GetArrayLength(context, arr_obj.into(), &mut l)};
+
+    if !ok {
+        if let Some(err) = report_es_ex(context) {
+            return Err(err);
+        }
+    }
+    Ok(l)
 }
 
-pub fn get_array_slot(_context: *mut JSContext, _arr_obj: HandleObject, _idx: u32, _ret_val: MutableHandleValue) {
-    panic!("NYI");
+/// get the Value of an index of an Array
+pub fn get_array_slot(context: *mut JSContext, arr_obj: HandleObject, idx: i32, ret_val: MutableHandleValue) -> Result<(), EsErrorInfo> {
+
+    // could use https://developer.mozilla.org/en-US/docs/Mozilla/Projects/SpiderMonkey/JSAPI_reference/JS_ValueToId
+    // but the glue is here so thats easier
+
+    let id = unsafe{int_to_jsid(idx)};
+    rooted!(in (context) let id_root = id);
+
+    let ok = unsafe{JS_GetPropertyById(context, arr_obj.into(), id_root.handle().into(), ret_val.into())};
+
+    if !ok {
+        if let Some(err) = report_es_ex(context) {
+            return Err(err);
+        }
+    }
+
+    Ok(())
 }
 
+/// convert an Array to a Vec<i32>
 pub fn to_i32_vec(context: *mut JSContext, obj: HandleValue) -> Vec<i32> {
     let converted = unsafe {
         Vec::<i32>::from_jsval(context, obj,
@@ -39,6 +74,7 @@ pub fn to_i32_vec(context: *mut JSContext, obj: HandleValue) -> Vec<i32> {
     vec_ref.to_vec()
 }
 
+/// convert a Vec<i32> to an Array
 pub fn to_i32_array<T>(context: *mut JSContext, obj: MutableHandleValue, vec: Vec<i32>) {
     unsafe {vec.to_jsval( context, obj)};
 }
@@ -46,8 +82,11 @@ pub fn to_i32_array<T>(context: *mut JSContext, obj: MutableHandleValue, vec: Ve
 #[cfg(test)]
 mod tests {
     use crate::es_utils::tests::test_with_sm_rt;
-    use crate::es_utils::{get_es_obj_prop_val, report_es_ex};
-    use crate::es_utils::arrays::{get_array_length, object_is_array};
+    use crate::es_utils::report_es_ex;
+    use crate::es_utils::arrays::{get_array_length, object_is_array, get_array_slot};
+    use crate::es_utils::objects::get_es_obj_prop_val;
+    use mozjs::jsval::JSVal;
+    use mozjs::jsval::UndefinedValue;
 
     #[test]
     fn test_is_array(){
@@ -69,16 +108,29 @@ mod tests {
                 }
             }
 
-            let arr_val = get_es_obj_prop_val(context, global_root.handle(), "test_is_array");
-            let arr_val2 = get_es_obj_prop_val(context, global_root.handle(), "test_is_array2");
-            rooted!(in (context) let arr_val_root = arr_val.to_object());
-            rooted!(in (context) let arr_val2_root = arr_val2.to_object());
+            rooted!(in (context) let mut arr_val_root = UndefinedValue());
+            rooted!(in (context) let mut arr_val2_root = UndefinedValue());
 
-            assert_eq!(true, object_is_array(context,  arr_val_root.handle()));
+            let _res = get_es_obj_prop_val(context, global_root.handle(), "test_is_array", arr_val_root.handle_mut());
+            let _res2 = get_es_obj_prop_val(context, global_root.handle(), "test_is_array2", arr_val2_root.handle_mut());
 
-            assert_eq!(4, get_array_length(context, arr_val_root.handle()));
+            rooted!(in (context) let arr_val_obj = arr_val_root.to_object());
+            rooted!(in (context) let arr_val2_obj = arr_val2_root.to_object());
 
-            assert_eq!(false, object_is_array(context,  arr_val2_root.handle()));
+            assert_eq!(true, object_is_array(context,  arr_val_obj.handle()));
+
+            let length_res = get_array_length(context, arr_val_obj.handle());
+            assert_eq!(4, length_res.ok().unwrap());
+
+            assert_eq!(false, object_is_array(context,  arr_val2_obj.handle()));
+
+            rooted!(in (context) let mut rval = UndefinedValue());
+            let res = get_array_slot(context, arr_val_obj.handle(), 2, rval.handle_mut());
+            if res.is_err() {
+                panic!(res.err().unwrap().message);
+            }
+            let val_3: JSVal = *rval;
+            assert_eq!(val_3.to_int32(), 3);
 
             true
         });